Loads of minor minor updates
Also momentum is fucked for the player so fix that

+++ source/essential/random.h
@@ -11,7 +11,7 @@
 #define ROUNDING 0.000001
 
 template<class T>
-float radians(T degrees) {
+double radians(T degrees) {
 	return degrees * (M_PI / 180.0);
 }
